Rename frå "api" til "handlers" for å vere tettare på terminologien til msw
