+++ src/mock/index.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw/browser';
-import { allHandlers } from './api';
+import { allHandlers } from './handlers';
 
 export const worker = setupWorker(...allHandlers);
